ENH: Move all the logic outside of ShrinkImage
The CLI ShrinkImage accepts an argument that specifies the size of its output. This
arguments was used to compute the shrink factors in the CLI, and the shrink factors
were then passed to itktubeShrinkWithBlendingImageFilter. This computation has
been moved to itktubeShrinkWithBlendingImageFilter. This allow all the logic to
be inside the filter, and therefore when wrapping the filter in Python, one has
access to this functionality.