ENH: TubeSpatialObjectToImageFilter is not limited to float
When not specifying some templated arguments of TubeSpatialObjectToImageFilter,
default arguments are used. These default arguments instantiate itk::Image
of type float. This limit the usage of this class to image of that type.
To allow wrapping this class in Python for other pixel types, all
templated parameters are manually specified when creating of object of
the class TubeSpatialObjectToImageFilter.